+++ tests/Unit/TwitchySharp.Core.Tests.Unit/Primitives/Test_CharityAmount.cs
@@ -9,7 +9,7 @@ public class Test_CharityAmount
     {
         public required int Value { get; set; }
         public required int DecimalPlaces { get; set; }
-        public required double Expected { get; set; }
+        public required decimal Expected { get; set; }
 
         public CharityAmount CharityAmount => new()
         {
@@ -22,7 +22,7 @@ public class Test_CharityAmount
         {
             Value = info.GetValue<int>(nameof(Value));
             DecimalPlaces = info.GetValue<int>(nameof(DecimalPlaces));
-            Expected = info.GetValue<double>(nameof(Expected));
+            Expected = info.GetValue<decimal>(nameof(Expected));
         }
         public void Serialize(IXunitSerializationInfo info)
         {
@@ -37,11 +37,11 @@ public class Test_CharityAmount
 
     public static IEnumerable<TheoryDataRow<TestCharityAmount>> MonetaryValueData =>
         [
-            new(new() { Value = 100, DecimalPlaces = 2, Expected = 1.00 }),
-            new(new() { Value = 250, DecimalPlaces = 0, Expected = 250.0 }),
-            new(new() { Value = 15, DecimalPlaces = 1, Expected = 1.5 }),
-            new(new() { Value = 2382, DecimalPlaces = 3, Expected = 2.382 }),
-            new(new() { Value = 5000000, DecimalPlaces = 6, Expected = 5.0 }),
+            new(new() { Value = 100, DecimalPlaces = 2, Expected = 1.00m }),
+            new(new() { Value = 250, DecimalPlaces = 0, Expected = 250.0m }),
+            new(new() { Value = 15, DecimalPlaces = 1, Expected = 1.5m }),
+            new(new() { Value = 2382, DecimalPlaces = 3, Expected = 2.382m }),
+            new(new() { Value = 5000000, DecimalPlaces = 6, Expected = 5.0m }),
         ];
 
     [Theory]
